Refactor to pull common code into new ConditionalTestDiscoverer
