+++ src/xunit.netcore.extensions/Discoverers/ConditionalFactDiscoverer.cs
@@ -5,8 +5,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
-using System.Text;
 using Xunit.Abstractions;
 using Xunit.Sdk;
 
@@ -26,95 +24,7 @@ namespace Xunit.NetCore.Extensions
         {
             string[] conditionMemberNames = factAttribute.GetConstructorArguments().FirstOrDefault() as string[];
             IEnumerable<IXunitTestCase> testCases = base.Discover(discoveryOptions, testMethod, factAttribute);
-            return DiscoverConditionalTestCases(discoveryOptions, _diagnosticMessageSink, testMethod, testCases, conditionMemberNames);
-        }
-
-        // This helper method evaluates the given condition member names for a given set of test cases.
-        // If any condition member evaluates to 'false', the test cases are marked to be skipped.
-        // The skip reason is the collection of all the condition members that evalated to 'false'.
-        internal static IEnumerable<IXunitTestCase> DiscoverConditionalTestCases(
-                                                        ITestFrameworkDiscoveryOptions discoveryOptions,
-                                                        IMessageSink diagnosticMessageSink,
-                                                        ITestMethod testMethod,
-                                                        IEnumerable<IXunitTestCase> testCases,
-                                                        IEnumerable<string> conditionMemberNames)
-        {
-            MethodInfo testMethodInfo = testMethod.Method.ToRuntimeMethod();
-            Type testMethodDeclaringType = testMethodInfo.DeclaringType;
-            List<string> falseConditions = new List<string>();
-
-            foreach (string entry in conditionMemberNames)
-            {
-                string conditionMemberName = entry;
-                string[] symbols = conditionMemberName.Split('.');
-                Type declaringType = testMethodDeclaringType;
-
-                if (symbols.Length == 2)
-                {
-                    conditionMemberName = symbols[1];
-                    ITypeInfo type = testMethod.TestClass.Class.Assembly.GetTypes(false).Where(t => t.Name.Contains(symbols[0])).FirstOrDefault();
-                    if (type != null)
-                    {
-                        declaringType = type.ToRuntimeType();
-                    }
-                }
-
-                MethodInfo conditionMethodInfo;
-                if (conditionMemberName == null ||
-                    (conditionMethodInfo = LookupConditionalMethod(declaringType, conditionMemberName)) == null)
-                {
-                    return new[] {
-                    new ExecutionErrorTestCase(
-                        diagnosticMessageSink,
-                        discoveryOptions.MethodDisplayOrDefault(),
-                        testMethod,
-                        GetFailedLookupString(conditionMemberName))
-                };
-                }
-
-                // In the case of multiple conditions, collect the results of all
-                // of them to produce a summary skip reason.
-                if (!(bool)conditionMethodInfo.Invoke(null, null))
-                {
-                    falseConditions.Add(conditionMemberName);
-                }
-            }
-
-            // Compose a summary of all conditions that returned false.
-            if (falseConditions.Count > 0)
-            {
-                string skippedReason = String.Format("Condition(s) not met: \"{0}\"", String.Join("\", \"", falseConditions));
-                return testCases.Select(tc => new SkippedTestCase(tc, skippedReason));
-            }
-
-            // No conditions returned false (including the absence of any conditions).
-            return testCases;
-        }
-
-        internal static string GetFailedLookupString(string name)
-        {
-            return
-                "An appropriate member \"" + name + "\" could not be found. " +
-                "The conditional method needs to be a static method or property on this or any ancestor type, " +
-                "of any visibility, accepting zero arguments, and having a return type of Boolean.";
-        }
-        
-        internal static MethodInfo LookupConditionalMethod(Type t, string name)
-        {
-            if (t == null || name == null)
-                return null;
-
-            TypeInfo ti = t.GetTypeInfo();
-
-            MethodInfo mi = ti.GetDeclaredMethod(name);
-            if (mi != null && mi.IsStatic && mi.GetParameters().Length == 0 && mi.ReturnType == typeof(bool))
-                return mi;
-
-            PropertyInfo pi = ti.GetDeclaredProperty(name);
-            if (pi != null && pi.PropertyType == typeof(bool) && pi.GetMethod != null && pi.GetMethod.IsStatic && pi.GetMethod.GetParameters().Length == 0)
-                return pi.GetMethod;
-
-            return LookupConditionalMethod(ti.BaseType, name);
+            return ConditionalTestDiscoverer.Discover(discoveryOptions, _diagnosticMessageSink, testMethod, testCases, conditionMemberNames);
         }
     }
 }
